refactor(suggestions): tighten the folder's seams and make downloads user-initiated only
Drops the onboarding warm-up: the enable chip and the settings Download
button are the explicit, user-initiated download paths. Folds the
availability probe into use-engine-view as private machinery (it had a
single consumer), renames the derivation module after its export the
way to-phrases is, and trims the pending flags from the hook's return —
pending is a status the bar renders, not a flag anyone polls.
