+++ src/app/layouts/app-shell.test.tsx
@@ -1,5 +1,4 @@
 import { AppProviders } from "@shared/providers/app-providers";
-import { stubProofreader, stubRewriter } from "@shared/testing/built-in-ai";
 import { createMemoryRouter, RouterProvider } from "react-router";
 import { describe, expect, test, vi } from "vitest";
 import { render } from "vitest-browser-react";
@@ -21,10 +20,7 @@ function renderAppShell() {
 }
 
 describe("AppShell", () => {
-  test("dismissing onboarding warms up the suggestion models", async () => {
-    const proofreader = stubProofreader();
-    const rewriter = stubRewriter();
-
+  test("onboarding shows on first visit and closes on Continue", async () => {
     const screen = await renderAppShell();
 
     await expect
@@ -33,10 +29,6 @@ describe("AppShell", () => {
 
     await screen.getByRole("button", { name: "Continue" }).click();
 
-    await vi.waitFor(() => {
-      expect(proofreader.create).toHaveBeenCalledOnce();
-      expect(rewriter.create).toHaveBeenCalledOnce();
-    });
     await expect
       .element(screen.getByRole("dialog", { name: "AAC Board AI" }))
       .not.toBeInTheDocument();
