+++ src/features/board/suggestions/suggestion-bar.tsx
@@ -3,7 +3,7 @@ import Chip from "@mui/material/Chip";
 import Stack from "@mui/material/Stack";
 import { SuggestionStatus } from "./suggestion-status";
 import { ToneSelector } from "./tone-selector";
-import type { SuggestionStatusView } from "./suggestion-status-view";
+import type { SuggestionStatusView } from "./derive-suggestion-status";
 
 export interface SuggestionBarProps {
   suggestions: string[];
